Added gmsh to requirements
